Massa v27 compatible (#1274)
* 1261 updagre get filtered sc output events for v27 (#1262)

* fix variable typo

* fix order of event parameter

* typos

* change op type id type

* update max gas to v27 values

* web-on-chain: improve and adjust for v27 (#1264)

* web-on-chain: reduce chunk size to 78000

* web-on-chain: add description

* roll back DefaultGasLimit

* rollback chunk size

* increase default max_gas for execute SC

* update task file

* clean code: remove if statements for v<26

+++ pkg/node/sendoperation/transaction/transaction.go
@@ -11,7 +11,7 @@ import (
 )
 
 const (
-	OpType      = uint64(0)
+	OpType      = 0
 	versionByte = byte(0)
 )
